Adds a layer with the events on top of a plot of an eeg_lst.

Adds a layer with the events on top of a plot of an eeg_lst.

annotate_events(data = NULL, alpha = 0.2)

Arguments

data
The data to be displayed in this layer. There are three options: * If NULL, the default, the events table is inherited from the plot data as specified in the call to ggplot(). * An events table will override the plot events table data.

alpha
new alpha level in 0,1.

Value

A ggplot layer.

See also