Convert an eeg_lst to a (base) data frame.

Convert an eeg_lst to a (base) data frame.

# S3 method for eeg_lst
as.data.frame(x, row.names = NULL, optional = FALSE, ..., .unit = "second")

Arguments

x

An eeg_lst object.

row.names

NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional

logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional. Note that all of R's base package as.data.frame() methods use optional only for column names treatment, basically with the meaning of data.frame(*, check.names = !optional). See also the make.names argument of the matrix method.

...

Additional arguments to be passed to or from other methods.

.unit

Unit for the .time column of the transformed object: "s" (default), "ms", "samples".

Value

A data.frame.