Dplyr functions for manipulating eeg_lst objects.

Manipulate the signal_tbl and the segments table of an eeg_lst.

# S3 method for eeg_lst
mutate(.data, ...)

# S3 method for eeg_lst
transmute(.data, ...)

# S3 method for eeg_lst
filter(.data, ...)

# S3 method for eeg_lst
summarise(.data, ...)

# S3 method for eeg_lst
group_by(.data, ..., add = FALSE, .drop = FALSE)

# S3 method for eeg_lst
ungroup(.data, ...)

# S3 method for eeg_lst
select(.data, ...)

# S3 method for eeg_lst
rename(.data, ...)

Arguments

.data
An eeg_lst.

...
Name-value pairs of expressions; see dplyr for more help.

Value

An eeg_lst object.

Details

Wrappers for dplyr's commands that act on different parts eggble objects. The following wrappers have been implemented for eeg_lst objects:

mutate() adds new variables and preserves existing ones. Variables that are a function of a channel are added to the signal_tbl table, and other variables are added to the segments table.

transmute() like mutate but drops non-used variables of the referred table, except for the obligatory columns starting with ..

filter(): finds segments/samples where conditions are true. Segments/samples where the condition evaluates to NA are dropped.

summarize() summarizes the channel of the signal_tbl table

group_by(): allows that operations would be performed "by group".

ungroup(): removes the grouping created by group_by.

select() keeps only the mentioned variables from the referred table, except for the obligatory columns starting with ..

rename(): keeps all variables.

These commands always return the entire eeg_lst so that they can be ' piped using magrittr 's pipe,

See also

Examples

# NOT RUN {
faces_segs %>% select(O1, O2, P7, P8)
# }