Downsample EEG data

Downsample a signal_tbl by a factor q, using an FIR or IIR filter. This is a wrapper for decimate from the signal package, see its documentation for details. Notice that the code of the signal package might be outdated. This function is used in plotting functions.

eeg_downsample(
  .data,
  .q = 2,
  .max_sample = NULL,
  .n = if (.ftype == "iir") 8 else 30,
  .ftype = "iir",
  .multiple_times = FALSE,
  ...
)

Arguments

.data

An eeg_lst object.

.q

integer factor(s) to downsample by.

.max_sample

Optionally, the (approximated) maximum sample number can be defined here, which is at least half of the total number of samples

.n

filter order used in the downsampling.

.ftype

filter type, "iir" or "fir"

.multiple_times

Indicates whether to factorize `q`` and apply the downsampling in steps.

...

Not in use.

Details

A factor q larger than 13 can result in NAs. To avoid this, the downsampling can be done in steps. For example, instead of setting q = 20, it is possible to set q = c(2,10).

See also