Re-reference a channel or group of channels.

Re-reference a channel or group of channels.

eeg_rereference(.data, ..., ref = NULL, na.rm = FALSE)

Arguments

.data
An eeg_lst object.

...
Channels to include. All the channels by default, but eye channels channels should be removed.

ref
Character vector of channels that will be averaged as the reference.

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Value

An eeg_lst with some channels re-referenced.

Details

Notice that this function will also rereference the eye electrodes unless excluded. See examples.

See also

Examples

# Re-reference all channels using the left mastoid excluding the eye electrodes.
data_faces_ERPs_M1 <- data_faces_ERPs %>% 
                      eeg_rereference(-EOGV, -EOGH, ref = "M1")

# Re-reference using the linked mastoids excluding the eye electrodes.
data_faces_ERPs_M1M2 <- data_faces_ERPs %>% 
                        eeg_rereference(-EOGV, -EOGH, ref = c("M1","M2"))