,

eeguana: Flexible Manipulation of EEG Data

Flexible manipulation of EEG data in R. It provides a data.table powered framework for manipulating EEG data with dplyr-based functions (e.g., `mutate`, `filter`, `summarize`) extended to a new class, eeg_lst, other EEG-specialized functions, and `ggplot` wrapper functions. The new class is inspired by tidyverse principles but it's not really "tidy" (due to space considerations), it's a list of (i) a wide data table (signal_tbl) that contains the signal amplitudes at every sample point of the EEG, (ii) an events data table with information about markers (or triggers), blinks and other exported information, and (iii) a long table with experimental information, such as participant number (.recording), conditions, etc. eeguana can do only basic pre-processing for now.

See also

Author

Maintainer: Bruno Nicenboim bruno.nicenboim@tilburguniversity.edu

Other contributors:

Kate Stone kate.stone@uni-potsdam.de [contributor]

Mark Fairbanks mark.t.fairbanks@gmail.com [contributor]