Dplyr functions for joining data frames to the segments of eeg_lst objects.

Join a data frames to the segments of an eeg_lst object, and modify the signal table accordingly (dropping rows when necessary).

# S3 method for eeg_lst
anti_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for eeg_lst
left_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for eeg_lst
semi_join(x, y, by = NULL, suffix = c(".x", ".y"),
  ...)

Arguments

x
An eeg_lst.

y
A data frame, tibble, or data.table.

by
a character vector of variables to join by. If NULL, the default, *_join() will do a natural join, using all variables with common names across the two tables. A message lists the variables so that you can check they're right (to suppress the message, simply explicitly list the variables that you want to join).

To join by different variables on x and y use a named vector. For example, by = c("a" = "b") will match x.a to y.b.

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

...
other parameters passed onto methods, for instance, na_matches to control how NA values are matched. See join.tbl_df for more.

Value

An eeg_lst object.

Details

Wrappers for dplyr's commands: The following wrappers have been implemented for eeg_lst objects:

left_join(): left-joins an external table to the segments table of the eeg_lst.

semi_join(): semi-joins an external table to the segments table of the eeg_lst.

anti_join(): anti-joins an external table to the segments table of the eeg_lst. These commands always return the entire eeg_lst so that they can be ' piped using magrittr 's pipe,

See also