Summarize multiple channels

These are scoped variants of summarize() that apply operations on a selection of channels

summarize_at_ch(.tbl, ...)

summarize_all_ch(.tbl, ...)

# S3 method for eeg_lst
summarize_at_ch(.tbl, .vars, .funs, ...)

# S3 method for eeg_lst
summarize_all_ch(.tbl, .funs, ...)

summarise_all_ch(.tbl, ...)

Arguments

.tbl
An eeg_lst object.

...
Additional arguments for the function calls in .funs. These are evaluated only once, with tidy dots support.

.vars
A list of channels generated by vars(), a character vector of channel names, or NULL.

.funs
List of function calls generated by funs(), or a character vector of function names, or simply a function.

Bare formulas are passed to rlang::as_function() to create purrr-style lambda functions. Note that these lambda prevent hybrid evaluation from happening and it is thus more efficient to supply functions like mean() directly rather than in a lambda-formula.

Details

summarize_at_ch allows for selecting the channels using the same name-based select_helpers just like with select().

summarize_all_ch summarizes all channels.

See also