,

Select independent components (or sources) to keep.

This function will transform the channels according to the independent components that are kept or removed.

eeg_ica_keep(.data, ...)

Arguments

.data

An eeg_ica_lst object

...

Components to keep from the mixing matrix of the ICA transformation. See dplyr::select and tidyselect::select_helpers for details.

See also

Examples

# For demonstration only, since ICA won't converge
# Suppressing an important warning:
suppressWarnings(data_faces_10_trials %>%
  eeg_ica(-EOGH, -EOGV, -M1, -M2, .method = fast_ICA, .config = list(maxit = 10))) %>%
  eeg_ica_keep(-ICA1)
#> # ICA is being done using fast_ICA...
#> # ICA finished in 9 iterations. With a tolerance of 0.830862. (Maximum expected tolerance: 1e-06.)
#> # ICA took 1.72 secs
#> # EEG data:
#> 
#> # Signal table:
#>        .id .sample       Fp1        Fpz         Fp2         F7        F3
#>     1:   1   15000 -32.24707  -4.388853  -2.8908265 -9.8411207 -7.104096
#>     2:   1   15001 -28.31158 -11.356616  -4.7101960  1.6875249 -9.017591
#>     3:   1   15002 -25.97296 -16.012272  -4.0300390 10.2069583 -7.914812
#>     4:   1   15003 -26.30960 -16.757605  -2.1077117 12.5648823 -6.851094
#>     5:   1   15004 -28.40190 -15.557377   0.2755778 10.8862797 -5.460543
#>    ---                                                                  
#> 43325:   1   58324 -17.83708 -13.980014  -7.9887605 -0.9040933 -7.881421
#> 43326:   1   58325 -18.59067 -13.694987  -7.4725155 -2.4067566 -6.845604
#> 43327:   1   58326 -17.92556 -14.090027  -8.5261421 -2.2981401 -7.172108
#> 43328:   1   58327 -16.32854 -14.932628  -9.8475955 -2.0971855 -7.866034
#> 43329:   1   58328 -15.00734 -15.317460 -10.6712208 -1.6170765 -8.344716
#>               Fz          F4        F8         FC5         FC1         FC2
#>     1:  2.514379  7.83178899 12.241794   0.1409402 -4.72093861  5.46559892
#>     2: -1.970082  0.09887244  7.065749   5.6998908 -6.15539217 -2.13532314
#>     3: -5.523369 -5.05490063  3.352800   2.4809057 -4.99216711 -8.00642504
#>     4: -6.461054 -5.69100212  2.488400  -2.1646820 -4.12776992 -9.57096489
#>     5: -5.571272 -3.37820001  4.904046  -0.7711080 -3.94558547 -8.00421205
#>    ---                                                                    
#> 43325: -7.375420 -2.60813266 -5.843741  -8.1282673 -1.24794556 -1.62617739
#> 43326: -6.531925 -0.80177138 -4.637263 -10.5398151 -0.23652468 -0.02911744
#> 43327: -6.173130 -0.35596788 -4.566375  -8.8559154 -0.05406192  0.05495427
#> 43328: -6.183953 -0.07888685 -3.615071  -3.9064973 -0.89529491 -0.34971759
#> 43329: -6.149312 -0.14947745 -3.049929   1.9170941 -2.19283876 -0.87235315
#>              FC6         M1           T7         C3          Cz         C4
#>     1: 22.687154 -1.8668146 -17.63149537 -9.5623914 -10.3045263 14.3780125
#>     2:  2.430781 -3.2960243  -7.51778011 -6.2148278 -10.8953177  6.2757094
#>     3: -9.640350 -3.7450671  -1.76144255  1.0528794 -10.4875653  0.7020062
#>     4: -9.249898 -3.7946916  -5.79388770  5.5393554 -11.5960161  0.3527392
#>     5: -2.711073 -3.2719884 -11.82030552  4.7051224 -13.1771777  2.1397569
#>    ---                                                                    
#> 43325: -7.361213  0.8050708  -0.63608403 -0.8986476  -3.5814171 -6.5431452
#> 43326: -4.954895  0.9832335  -0.65014938  0.1947319  -1.8897396 -4.4015992
#> 43327: -3.739712  1.2051768  -0.04028933  0.7194492  -0.8786211 -3.6439259
#> 43328: -2.872486  1.4103985   0.50926713 -1.1124471  -1.1049968 -3.7887957
#> 43329: -2.939587  1.1627738   0.79255033 -4.0596354  -2.2833442 -4.1566411
#>                T8         M2         CP5         CP1       CP2        CP6
#>     1: -14.647893  1.8668146 -6.80513655 -15.6472291 -2.914130  16.182018
#>     2:  -2.135390  3.2960243 -0.99800861 -13.6671606 -5.723453  13.354459
#>     3:  10.005640  3.7450671  5.93292034  -9.8446659 -6.053359  13.164176
#>     4:  10.019100  3.7946916 11.28958135  -8.4099914 -5.964303  15.527580
#>     5:   5.070975  3.2719884 12.40333312  -9.6694562 -6.540058  15.316073
#>    ---                                                                   
#> 43325:  -2.149223 -0.8050708  0.45155474   0.4148388 -7.043711 -11.471771
#> 43326:  -1.312486 -0.9832335 -0.09866453   1.5243880 -4.998291  -8.014661
#> 43327:   1.739134 -1.2051768  0.26993142   2.1559323 -4.143072  -6.142017
#> 43328:   6.279146 -1.4103985 -0.63580457   1.3639733 -4.455968  -5.345660
#> 43329:   7.930107 -1.1627738 -1.50936597  -0.2453353 -5.634644  -5.404735
#>                 P7          P3         Pz        P4        P8        POz
#>     1: -11.5150361 -10.7997954  -9.424803  1.190774 -6.215506 -14.668043
#>     2:  -0.7773977  -5.6507512  -9.447899 -1.323730 -4.387737 -11.235329
#>     3:   7.8026920  -0.1001299  -8.137937 -1.740233 -1.562850  -7.880214
#>     4:  10.9673748   2.5560204  -8.346501 -1.458035 -1.141949  -7.667208
#>     5:   9.3238525   1.2309271 -10.090177 -2.509041 -3.709198 -10.463789
#>    ---                                                                  
#> 43325:   1.9540273  -2.5587326  -4.652872 -8.767808 -9.063308  -6.105534
#> 43326:   0.2625960  -2.9669293  -3.368259 -6.577292 -6.969477  -4.969602
#> 43327:  -0.3532735  -2.6532892  -2.342847 -5.037537 -4.535009  -3.328878
#> 43328:  -1.9555808  -3.9237362  -2.572404 -4.250655 -2.419578  -2.831196
#> 43329:  -2.3339417  -5.4427013  -3.830606 -4.577994 -2.006789  -3.411969
#>                 O1         Oz         O2       EOGV       EOGH
#>     1: -19.3910830 -19.819019 -15.886975 -205.85693 164.320831
#>     2:  -9.7486124 -11.502271 -10.237117 -214.15259 158.389755
#>     3:  -2.3408198  -5.231102  -5.646818 -220.57965 157.559784
#>     4:  -0.5246842  -4.934055  -5.750961 -218.24934 162.165161
#>     5:  -4.4480029 -10.489347 -10.643831 -217.16670 168.559723
#>    ---                                                        
#> 43325:  -1.7625430  -7.210809  -8.796753   18.41424   6.434383
#> 43326:  -2.5045389  -6.436767  -7.161250   19.22052   7.837689
#> 43327:  -2.5131521  -4.320360  -4.058998   19.74028   8.659687
#> 43328:  -3.2994951  -3.106040  -1.656144   19.20309   8.609533
#> 43329:  -3.3750895  -2.690126  -1.128599   18.72752   8.829977
#> 
#> ## Events table:
#>      .id        .type .description .initial .final .channel
#>   1:   1 Bad Interval  Bad Min-Max    15000  15128      FC6
#>   2:   1 Bad Interval  Bad Min-Max    15000  15146       F8
#>   3:   1 Bad Interval  Bad Min-Max    15000  15142      Fp1
#>   4:   1 Bad Interval  Bad Min-Max    15000  15132      Fp2
#>   5:   1 Bad Interval  Bad Min-Max    15000  15140      Fpz
#>  ---                                                       
#> 540:   1 Bad Interval  Bad Min-Max    55058  55400      Fpz
#> 541:   1 Bad Interval  Bad Min-Max    55077  55314       F8
#> 542:   1     Stimulus         s130    56380  56380     <NA>
#> 543:   1     Stimulus          s70    56750  56750     <NA>
#> 544:   1     Stimulus         s130    58328  58328     <NA>
#> 
#> # ICA:
#> # Component_names:ICA2, ICA3, ICA4, ICA5, ICA6, ICA7, ICA8, ICA9, ICA10, ICA11, ICA12, ICA13, ICA14, ICA15, ICA16, ICA17, ICA18, ICA19, ICA20, ICA21, ICA22, ICA23, ICA24, ICA25, ICA26, ICA27, ICA28, ICA29, ICA30
#> # Channels_used: Fp1, Fpz, Fp2, F7, F3, Fz, F4, F8, FC5, FC1, FC2, FC6, T7, C3, Cz, C4, T8, CP5, CP1, CP2, CP6, P7, P3, Pz, P4, P8, POz, O1, Oz, O2
#> 
#> # Segments table:
#>    .id .recording segment
#> 1:   1 faces.vhdr       1