Select independent components (or sources) to keep.

This function will transform the channels according to the independent components that are kept or removed.

eeg_ica_keep(.data, ...)

Arguments

.data

An eeg_ica_lst object

...

Components to keep from the mixing matrix of the ICA transformation. See dplyr::select and tidyselect::select_helpers for details.

See also

Examples

# For demonstration only, since ICA won't converge
# Suppressing an important warning:
suppressWarnings(data_faces_10_trials %>%
  eeg_ica(-EOGH, -EOGV, -M1, -M2, .method = fast_ICA, .config = list(maxit = 10))) %>%
  eeg_ica_keep(-ICA1)
#> # ICA is being done using fast_ICA...
#> # ICA finished in 9 iterations. With a tolerance of 0.791687. (Maximum expected tolerance: 1e-06.)
#> # ICA took 1.54 secs
#> # EEG data:
#> 
#> # Signal table:
#>        .id .sample       Fp1        Fpz        Fp2         F7        F3
#>     1:   1   15000 -29.58899  -1.629751  -4.681317 -0.8498115 -9.629681
#>     2:   1   15001 -28.74177 -11.144808  -5.747874  1.8643604 -9.371787
#>     3:   1   15002 -28.46713 -17.479296  -4.641227  4.5554025 -6.900461
#>     4:   1   15003 -29.43772 -18.570272  -2.961469  5.7550023 -5.734929
#>     5:   1   15004 -31.30729 -16.932404  -1.150009  5.4932385 -4.902847
#>    ---                                                                 
#> 43325:   1   58324 -19.68453 -14.066859 -10.584809 -1.1369161 -9.033971
#> 43326:   1   58325 -20.99914 -13.884950 -10.711134 -3.1359212 -8.223015
#> 43327:   1   58326 -20.28918 -14.426307 -11.386807 -3.6322681 -8.251204
#> 43328:   1   58327 -19.67281 -15.698081 -13.257030 -5.1171492 -8.857478
#> 43329:   1   58328 -19.48437 -16.621795 -14.615079 -6.6773446 -9.166581
#>               Fz         F4         F8          FC5        FC1        FC2
#>     1:  2.159375  2.8939415   3.451874  5.823980285 -6.5520579  1.1834461
#>     2: -1.921382 -0.8191964   4.829885  6.457585897 -5.8502712 -2.5175075
#>     3: -5.219464 -3.3721942   5.322716  0.002239164 -3.2851761 -5.8792374
#>     4: -6.169556 -3.7722709   4.731402 -5.171342997 -2.1161658 -7.0862971
#>     5: -5.405870 -2.4952326   5.251993 -2.796014691 -2.2014255 -6.3077698
#>    ---                                                                   
#> 43325: -7.828387 -4.7561812 -10.218273 -7.276623608 -0.8561749 -2.4604344
#> 43326: -7.033624 -3.3094335  -9.940499 -9.645824171  0.3734350 -0.9466773
#> 43327: -6.562527 -2.3034972  -8.900333 -8.391109167  0.6734266 -0.4927169
#> 43328: -6.559997 -1.8528168  -8.037495 -4.048797989  0.3703288 -0.4232950
#> 43329: -6.509962 -1.6047695  -7.311301  0.962750594 -0.3041574 -0.3086520
#>              FC6         M1         T7         C3          Cz        C4
#>     1: 16.731038 -1.8668146 -15.918504 -7.0724648 -12.5737023  9.564594
#>     2:  1.374523 -3.2960243  -6.427453 -5.2727412 -10.4391575  5.834759
#>     3: -7.529177 -3.7450671  -1.046808  0.9982652  -8.2374548  3.122605
#>     4: -6.784168 -3.7946916  -4.857739  5.5407818  -8.9138658  3.188974
#>     5: -1.425748 -3.2719884 -10.315257  5.3481649 -10.7761208  4.072886
#>    ---                                                                 
#> 43325: -9.372635  0.8050708   1.920542  1.2661724  -2.7315164 -7.642484
#> 43326: -7.319824  0.9832335   2.500864  2.7954194  -0.6838936 -5.554598
#> 43327: -5.526006  1.2051768   2.736595  2.9244116   0.4270585 -4.331971
#> 43328: -4.355729  1.4103985   3.856406  1.3716615   0.9663326 -3.939269
#> 43329: -3.927050  1.1627738   4.714345 -1.3326625   0.6695380 -3.608703
#>                 T8         M2        CP5        CP1       CP2        CP6
#>     1: -20.7318865  1.8668146  1.4444320 -15.644722 -7.385550   7.174543
#>     2:  -5.1022621  3.2960243  0.7590248 -12.738200 -5.920451  12.361322
#>     3:   8.9750748  3.7450671  3.4912509  -8.267749 -3.465405  17.399655
#>     4:   8.6982246  3.7946916  8.4847591  -6.457050 -2.860524  20.550100
#>     5:   2.2010851  3.2719884 11.3320313  -7.466143 -4.148469  18.703993
#>    ---                                                                  
#> 43325:  -7.6943869 -0.8050708  4.0993459   2.553478 -7.072016 -13.217566
#> 43326:  -8.3007178 -0.9832335  4.2105939   4.277644 -4.912235  -9.890143
#> 43327:  -4.4082975 -1.2051768  3.6701160   4.719615 -3.775419  -7.246618
#> 43328:  -0.9698571 -1.4103985  2.5943755   4.719793 -3.357373  -5.492023
#> 43329:  -0.3086748 -1.1627738  1.3130335   3.956778 -3.616558  -4.272781
#>               P7         P3          Pz         P4          P8         POz
#>     1: 0.8365279 -4.0135274 -11.0581857 -5.1400544 -12.8135566 -14.0337438
#>     2: 2.2365958 -3.3328164  -8.4970566 -1.6087619  -4.9973726  -9.9047601
#>     3: 4.8214569 -0.6252317  -5.4666126  1.9202202   1.7520404  -6.0496413
#>     4: 7.6247134  2.1014850  -5.0730881  2.9430906   2.8149459  -5.3748072
#>     5: 8.6937332  2.4494240  -6.8055540  0.8917466  -0.8956072  -7.7103284
#>    ---                                                                    
#> 43325: 8.3936345  2.6128126  -2.2346191 -8.7543844  -9.9809543  -2.9571233
#> 43326: 7.9895370  3.3279750  -0.2594183 -6.3837870  -7.8676093  -1.0048694
#> 43327: 5.9254391  2.6795551   0.6330739 -4.4520769  -4.9007436   0.2860196
#> 43328: 4.4177894  2.0112640   1.5164032 -2.6232189  -1.9491771   1.7684303
#> 43329: 3.7976221  0.9566357   1.5071650 -1.6422010  -0.4592636   2.2342026
#>                O1          Oz         O2       EOGV       EOGH
#>     1: -12.161682 -15.7416151 -15.885955 -205.85693 164.320831
#>     2:  -7.407219  -9.8191066  -9.443720 -214.15259 158.389755
#>     3:  -3.091970  -5.0521370  -4.314383 -220.57965 157.559784
#>     4:  -1.297724  -4.6184535  -4.077186 -218.24934 162.165161
#>     5:  -3.543559  -9.1206285  -8.724964 -217.16670 168.559723
#>    ---                                                        
#> 43325:   2.855095  -3.6890217  -6.740793   18.41424   6.434383
#> 43326:   3.211449  -2.0679294  -4.588961   19.22052   7.837689
#> 43327:   2.331677  -0.5396898  -1.704055   19.74028   8.659687
#> 43328:   2.032238   1.2523359   1.375691   19.20309   8.609533
#> 43329:   2.260213   2.1714856   2.642352   18.72752   8.829977
#> 
#> ## Events table:
#>      .id        .type .description .initial .final .channel
#>   1:   1 Bad Interval  Bad Min-Max    15000  15128      FC6
#>   2:   1 Bad Interval  Bad Min-Max    15000  15146       F8
#>   3:   1 Bad Interval  Bad Min-Max    15000  15142      Fp1
#>   4:   1 Bad Interval  Bad Min-Max    15000  15132      Fp2
#>   5:   1 Bad Interval  Bad Min-Max    15000  15140      Fpz
#>  ---                                                       
#> 540:   1 Bad Interval  Bad Min-Max    55058  55400      Fpz
#> 541:   1 Bad Interval  Bad Min-Max    55077  55314       F8
#> 542:   1     Stimulus         s130    56380  56380     <NA>
#> 543:   1     Stimulus          s70    56750  56750     <NA>
#> 544:   1     Stimulus         s130    58328  58328     <NA>
#> 
#> # ICA:
#> # Component_names:ICA2, ICA3, ICA4, ICA5, ICA6, ICA7, ICA8, ICA9, ICA10, ICA11, ICA12, ICA13, ICA14, ICA15, ICA16, ICA17, ICA18, ICA19, ICA20, ICA21, ICA22, ICA23, ICA24, ICA25, ICA26, ICA27, ICA28, ICA29, ICA30
#> # Channels_used: Fp1, Fpz, Fp2, F7, F3, Fz, F4, F8, FC5, FC1, FC2, FC6, T7, C3, Cz, C4, T8, CP5, CP1, CP2, CP6, P7, P3, Pz, P4, P8, POz, O1, Oz, O2
#> 
#> # Segments table:
#>    .id .recording segment
#> 1:   1 faces.vhdr       1