,

Select independent components (or sources) to keep.

This function will transform the channels according to the independent components that are kept or removed.

eeg_ica_keep(.data, ...)

Arguments

.data

An eeg_ica_lst object

...

Components to keep from the mixing matrix of the ICA transformation. See dplyr::select and tidyselect::select_helpers for details.

See also

Examples

# For demonstration only, since ICA won't converge
# Suppressing an important warning:
suppressWarnings(data_faces_10_trials %>%
  eeg_ica(-EOGH, -EOGV, -M1, -M2, .method = fast_ICA, .config = list(maxit = 10))) %>%
  eeg_ica_keep(-ICA1)
#> # ICA is being done using fast_ICA...
#> # ICA finished in 9 iterations. With a tolerance of 0.982379. (Maximum expected tolerance: 1e-06.)
#> # ICA took 1.66 secs
#> # EEG data:
#> 
#> # Signal table:
#>        .id .sample       Fp1        Fpz         Fp2        F7         F3
#>     1:   1   15000 -35.23106  -5.001963  -8.3947727 -6.204281 -11.595607
#>     2:   1   15001 -34.97236 -16.629889 -10.9635721 -2.098308 -13.533216
#>     3:   1   15002 -29.17531 -18.954975  -5.1305434  6.116569  -8.690620
#>     4:   1   15003 -29.07631 -19.439704  -2.6794897  8.532879  -7.245339
#>     5:   1   15004 -29.87777 -16.738663   0.2381033  9.047461  -5.621803
#>    ---                                                                  
#> 43325:   1   58324 -25.30661 -20.300519 -16.7830338 -4.453907 -13.907997
#> 43326:   1   58325 -23.18102 -17.273351 -13.7316492 -3.348937 -11.223917
#> 43327:   1   58326 -22.74004 -18.075858 -14.6844867 -4.013707 -11.413087
#> 43328:   1   58327 -19.54956 -17.465003 -14.2568537 -2.790159 -10.930599
#> 43329:   1   58328 -17.27384 -17.041721 -13.8583684 -1.943942 -10.563915
#>                 Fz        F4        F8         FC5       FC1       FC2
#>     1:   0.6607836  1.861829  3.236513   3.2941628 -6.615540  1.050703
#>     2:  -4.9259825 -3.136618  6.162562   3.4582602 -8.664602 -4.985666
#>     3:  -6.4665816 -4.242518  6.976669  -0.3250615 -5.634233 -7.610457
#>     4:  -7.1902380 -4.492219  6.533916  -4.9984748 -4.585168 -8.864788
#>     5:  -5.8546904 -2.744684  6.853297  -2.0443784 -4.115778 -7.568966
#>    ---                                                                
#> 43325: -11.2344280 -7.674347 -8.699674 -10.2360779 -4.632083 -6.045850
#> 43326:  -9.1002529 -5.100776 -8.497519 -10.8947770 -2.573996 -3.606324
#> 43327:  -8.7805159 -4.220677 -7.410503  -9.7786476 -2.386075 -3.245875
#> 43328:  -7.9864957 -3.082966 -6.290955  -4.1640904 -2.538549 -2.855093
#> 43329:  -7.4289331 -2.396647 -5.192715   1.8686815 -3.371692 -2.727256
#>               FC6         M1           T7         C3         Cz        C4
#>     1:  17.028412 -1.8668146 -15.38879467 -6.6652328 -11.804686  9.951918
#>     2:   1.153829 -3.2960243  -9.57273159 -6.9056482 -12.711369  4.364950
#>     3:  -7.749455 -3.7450671  -3.94164546 -0.6756449 -10.846658  1.716962
#>     4:  -7.090138 -3.7946916  -7.99479446  3.7191368 -11.825012  1.671923
#>     5:  -1.660540 -3.2719884 -12.78325163  3.8844168 -13.236347  2.910999
#>    ---                                                                   
#> 43325: -10.095567  0.8050708  -2.85678524 -1.2216581  -6.144579 -9.872466
#> 43326:  -7.914218  0.9832335  -1.34545265  0.7396955  -3.699203 -7.390379
#> 43327:  -6.131218  1.2051768  -1.19722422  0.8481265  -2.673317 -6.274475
#> 43328:  -4.926447  1.4103985  -0.00115807 -0.7550044  -2.377183 -5.891780
#> 43329:  -4.530597  1.1627738   0.52774570 -3.7120955  -3.195342 -5.752211
#>                 T8         M2        CP5         CP1       CP2        CP6
#>     1: -22.4061953  1.8668146  1.3622101 -14.2671189 -6.078327   8.886011
#>     2:  -2.3230064  3.2960243 -0.5412561 -14.5789858 -6.892780  12.069372
#>     3:  13.4639870  3.7450671  2.6165474 -10.9567391 -5.257108  15.776713
#>     4:  13.8041298  3.7946916  7.6064097  -9.5538263 -4.972736  18.532368
#>     5:   6.8669470  3.2719884 10.7588364 -10.1600782 -6.005135  16.816913
#>    ---                                                                   
#> 43325:  -3.8731220 -0.8050708  2.2113516  -0.6106017 -9.146990 -14.432072
#> 43326:  -4.2345583 -0.9832335  2.8562596   1.2589761 -6.979903 -11.494469
#> 43327:  -0.2570553 -1.2051768  2.2735597   1.6384732 -5.844129  -8.856988
#> 43328:   4.2573032 -1.4103985  1.3957594   1.1577437 -5.846626  -7.782093
#> 43329:   6.2771489 -1.1627738  0.1589893  -0.3391156 -6.696724  -7.330554
#>               P7          P3         Pz          P4          P8        POz
#>     1: 1.7065266 -2.18756084  -8.724875  -2.9440748 -11.1700951 -11.753951
#>     2: 0.3705809 -4.57886248  -9.759971  -1.9261185  -5.0177678 -10.954280
#>     3: 2.7142193 -3.30440588  -8.628938  -0.2429782   0.2996102  -9.000150
#>     4: 5.2553919 -1.07535845  -8.870885   0.2526390   0.9723928  -8.943345
#>     5: 6.7647676 -0.45065436 -10.332689  -1.6864740  -2.6950182 -11.055771
#>    ---                                                                    
#> 43325: 5.2490670  0.11369994  -4.974463 -10.1960983 -10.6670991  -5.349628
#> 43326: 5.2786140  0.61896439  -3.391607  -8.3948034  -9.0885315  -3.854860
#> 43327: 3.1532539 -0.01856089  -2.480387  -6.4120036  -6.1196643  -2.547984
#> 43328: 1.5005805 -1.44741032  -2.597536  -5.5249534  -3.9050595  -2.064760
#> 43329: 0.4775100 -3.45573391  -3.821158  -5.5930376  -3.1971314  -2.783065
#>                 O1         Oz          O2       EOGV       EOGH
#>     1: -11.0706293 -14.281580 -14.0678457 -205.85693 164.320831
#>     2:  -9.0619059 -10.589281  -9.3517335 -214.15259 158.389755
#>     3:  -5.6161607  -7.200621  -6.0380369 -220.57965 157.559784
#>     4:  -4.2109384  -7.198806  -6.2651740 -218.24934 162.165161
#>     5:  -6.1546933 -11.576113 -10.9425073 -217.16670 168.559723
#>    ---                                                         
#> 43325:   0.3512628  -5.141798  -7.2219530   18.41424   6.434383
#> 43326:   0.6914643  -3.923766  -5.8078363   19.22052   7.837689
#> 43327:  -0.2734270  -2.402766  -2.8558872   19.74028   8.659687
#> 43328:  -1.0893935  -1.359743  -0.7126990   19.20309   8.609533
#> 43329:  -1.5724812  -1.297205  -0.4095308   18.72752   8.829977
#> 
#> ## Events table:
#>      .id        .type .description .initial .final .channel
#>   1:   1 Bad Interval  Bad Min-Max    15000  15128      FC6
#>   2:   1 Bad Interval  Bad Min-Max    15000  15146       F8
#>   3:   1 Bad Interval  Bad Min-Max    15000  15142      Fp1
#>   4:   1 Bad Interval  Bad Min-Max    15000  15132      Fp2
#>   5:   1 Bad Interval  Bad Min-Max    15000  15140      Fpz
#>  ---                                                       
#> 540:   1 Bad Interval  Bad Min-Max    55058  55400      Fpz
#> 541:   1 Bad Interval  Bad Min-Max    55077  55314       F8
#> 542:   1     Stimulus         s130    56380  56380     <NA>
#> 543:   1     Stimulus          s70    56750  56750     <NA>
#> 544:   1     Stimulus         s130    58328  58328     <NA>
#> 
#> # ICA:
#> # Component_names:ICA2, ICA3, ICA4, ICA5, ICA6, ICA7, ICA8, ICA9, ICA10, ICA11, ICA12, ICA13, ICA14, ICA15, ICA16, ICA17, ICA18, ICA19, ICA20, ICA21, ICA22, ICA23, ICA24, ICA25, ICA26, ICA27, ICA28, ICA29, ICA30
#> # Channels_used: Fp1, Fpz, Fp2, F7, F3, Fz, F4, F8, FC5, FC1, FC2, FC6, T7, C3, Cz, C4, T8, CP5, CP1, CP2, CP6, P7, P3, Pz, P4, P8, POz, O1, Oz, O2
#> 
#> # Segments table:
#>    .id .recording segment
#> 1:   1 faces.vhdr       1