Select independent components (or sources) to keep.

This function will transform the channels according to the independent components that are kept or removed.

eeg_ica_keep(.data, ...)

Arguments

.data

An eeg_ica_lst object

...

Components to keep from the mixing matrix of the ICA transformation. See dplyr::select and tidyselect::select_helpers for details.

See also

Examples

# For demonstration only, since ICA won't converge
# Suppressing an important warning:
suppressWarnings(data_faces_10_trials %>%
  eeg_ica(-EOGH, -EOGV, -M1, -M2, .method = fast_ICA, .config = list(maxit = 10))) %>%
  eeg_ica_keep(-ICA1)
#> # ICA is being done using fast_ICA...
#> # ICA finished in 9 iterations. With a tolerance of 0.943982. (Maximum expected tolerance: 1e-06.)
#> # ICA took 1.56 secs
#> # EEG data:
#> 
#> # Signal table:
#>        .id .sample       Fp1       Fpz        Fp2          F7         F3
#>     1:   1   15000 -51.07497 -11.70108 -16.327657 -15.1524406 -17.529135
#>     2:   1   15001 -44.57045 -18.86525 -14.364297  -7.0420332 -15.565450
#>     3:   1   15002 -38.31383 -22.26104  -8.780833   1.6746861 -12.826249
#>     4:   1   15003 -36.96851 -22.30115  -5.817727   4.6923436 -10.897512
#>     5:   1   15004 -38.79355 -20.81939  -4.708748   3.7389281  -9.002341
#>    ---                                                                  
#> 43325:   1   58324 -18.25704 -13.90643 -11.103479  -0.2249610  -7.083544
#> 43326:   1   58325 -18.58288 -13.18646 -10.140707  -0.6816825  -6.721738
#> 43327:   1   58326 -17.71686 -13.62653 -10.604267  -0.9394603  -6.837867
#> 43328:   1   58327 -16.25611 -14.51485 -11.544011  -0.7394312  -7.944408
#> 43329:   1   58328 -14.79615 -14.96696 -12.136107  -0.6173644  -8.072177
#>               Fz         F4        F8       FC5         FC1        FC2
#>     1: -3.438075  0.4140791  6.530500 -5.895432 -11.1723339  -1.755580
#>     2: -6.375421 -2.2401119  9.172988 -2.654511 -10.9399276  -5.811707
#>     3: -9.040265 -5.8708743  7.798009 -5.253426  -8.3063280  -9.184254
#>     4: -9.455341 -5.9884251  7.139485 -9.219269  -6.8928990 -10.237546
#>     5: -8.245976 -3.5728026  8.828447 -7.283152  -6.7140327  -9.271077
#>    ---                                                                
#> 43325: -6.917161 -2.2694763 -6.773515 -7.721068  -1.4927515  -2.917165
#> 43326: -6.274839 -1.5150223 -7.145392 -9.278459  -0.5230147  -1.577636
#> 43327: -5.850906 -0.7054654 -6.349092 -7.849337  -0.1958005  -1.090716
#> 43328: -6.063916 -0.7858177 -5.611796 -2.900365  -1.1052561  -1.439254
#> 43329: -5.901050 -0.3765934 -4.300869  2.664401  -2.2811981  -1.674616
#>               FC6         M1          T7         C3         Cz         C4
#>     1: 20.1498307 -1.8668146 -18.4839768 -10.225165 -14.729358  9.6098996
#>     2:  4.8450775 -3.2960243 -10.4835194  -8.867753 -14.029867  5.5602433
#>     3: -7.1085328 -3.7450671  -5.3356999  -2.993491 -12.528402  0.7321580
#>     4: -6.6408204 -3.7946916  -9.1718032   1.700539 -13.269207  0.7053233
#>     5:  0.1998184 -3.2719884 -14.6319649   1.943993 -14.918649  2.7188304
#>    ---                                                                   
#> 43325: -6.2762524  0.8050708   0.1977102   1.045420  -3.851899 -5.8873774
#> 43326: -5.3250388  0.9832335   0.5968996   2.247225  -2.203334 -4.7327034
#> 43327: -3.8168513  1.2051768   0.9405663   2.404279  -1.070713 -3.7159796
#> 43328: -3.4267029  1.4103985   1.4085523   0.261764  -1.330106 -4.2174328
#> 43329: -2.9819083  1.1627738   1.4878630  -2.851058  -2.415041 -4.1894864
#>                 T8         M2       CP5         CP1       CP2        CP6
#>     1: -19.4444419  1.8668146 -1.668585 -17.6625738 -6.365628   7.872140
#>     2:   0.1143215  3.2960243 -2.910276 -17.1461736 -6.543018  12.014473
#>     3:  13.9162920  3.7450671  1.972783 -11.9317293 -5.464882  13.808035
#>     4:  14.0974374  3.7946916  7.147960 -10.2922939 -5.166558  16.703802
#>     5:   8.7646846  3.2719884  8.888559 -12.2024675 -6.177914  16.417291
#>    ---                                                                  
#> 43325:  -2.4219963 -0.8050708  1.216645  -1.2624086 -7.859922 -11.316884
#> 43326:  -3.1656357 -0.9832335  2.105927   0.7472508 -6.142301  -9.324525
#> 43327:   0.4800488 -1.2051768  1.825512   1.4236069 -4.993792  -6.915771
#> 43328:   4.7154180 -1.4103985  1.128919   1.0345940 -5.285886  -6.538538
#> 43329:   7.0589395 -1.1627738 -0.416367  -0.7761791 -6.238511  -5.941878
#>                P7         P3         Pz          P4         P8        POz
#>     1:  1.4143059 -4.1251928  -9.734789 -2.41476217 -13.628366 -12.816969
#>     2:  0.2587595 -6.8470533 -10.779621 -1.28501305  -6.900272 -12.290847
#>     3:  4.4394831 -2.5800452  -8.286668 -0.38313940  -2.052704  -8.469405
#>     4:  6.8964263 -0.2655448  -8.470914  0.08895061  -1.124122  -8.357138
#>     5:  6.2879643 -1.7392224 -10.989234 -1.31788239  -3.934342 -11.770216
#>    ---                                                                   
#> 43325:  3.4859118 -3.3789007  -6.396411 -9.29185622  -9.455424  -7.598309
#> 43326:  3.9348945 -1.8258809  -4.401436 -7.76387628  -8.190579  -5.420277
#> 43327:  2.2127142 -2.0623926  -3.283379 -5.87907691  -5.334966  -3.864080
#> 43328:  0.9332668 -2.7504177  -3.108894 -5.18234685  -3.422085  -2.910360
#> 43329: -0.5478314 -5.0586817  -4.520117 -5.17639550  -2.571064  -3.818430
#>                 O1         Oz         O2       EOGV       EOGH
#>     1: -14.0408718 -16.557696 -14.846413 -205.85693 164.320831
#>     2: -11.7512702 -12.873728 -10.533853 -214.15259 158.389755
#>     3:  -5.9436675  -7.893801  -6.478136 -220.57965 157.559784
#>     4:  -4.3389941  -7.707863  -6.618215 -218.24934 162.165161
#>     5:  -7.9855984 -12.881756 -11.311308 -217.16670 168.559723
#>    ---                                                        
#> 43325:  -1.7108365  -6.746146  -8.448609   18.41424   6.434383
#> 43326:  -0.7697441  -5.009057  -6.591424   19.22052   7.837689
#> 43327:  -1.3832336  -3.318820  -3.636142   19.74028   8.659687
#> 43328:  -1.7981571  -1.961088  -1.235731   19.20309   8.609533
#> 43329:  -2.5989423  -1.994960  -0.836466   18.72752   8.829977
#> 
#> ## Events table:
#>      .id        .type .description .initial .final .channel
#>   1:   1 Bad Interval  Bad Min-Max    15000  15128      FC6
#>   2:   1 Bad Interval  Bad Min-Max    15000  15146       F8
#>   3:   1 Bad Interval  Bad Min-Max    15000  15142      Fp1
#>   4:   1 Bad Interval  Bad Min-Max    15000  15132      Fp2
#>   5:   1 Bad Interval  Bad Min-Max    15000  15140      Fpz
#>  ---                                                       
#> 540:   1 Bad Interval  Bad Min-Max    55058  55400      Fpz
#> 541:   1 Bad Interval  Bad Min-Max    55077  55314       F8
#> 542:   1     Stimulus         s130    56380  56380     <NA>
#> 543:   1     Stimulus          s70    56750  56750     <NA>
#> 544:   1     Stimulus         s130    58328  58328     <NA>
#> 
#> # ICA:
#> # Component_names:ICA2, ICA3, ICA4, ICA5, ICA6, ICA7, ICA8, ICA9, ICA10, ICA11, ICA12, ICA13, ICA14, ICA15, ICA16, ICA17, ICA18, ICA19, ICA20, ICA21, ICA22, ICA23, ICA24, ICA25, ICA26, ICA27, ICA28, ICA29, ICA30
#> # Channels_used: Fp1, Fpz, Fp2, F7, F3, Fz, F4, F8, FC5, FC1, FC2, FC6, T7, C3, Cz, C4, T8, CP5, CP1, CP2, CP6, P7, P3, Pz, P4, P8, POz, O1, Oz, O2
#> 
#> # Segments table:
#>    .id .recording segment
#> 1:   1 faces.vhdr       1