Select independent components (or sources) to keep.

This function will transform the channels according to the independent components that are kept or removed.

eeg_ica_keep(.data, ...)

Arguments

.data

An eeg_ica_lst object

...

Components to keep from the mixing matrix of the ICA transformation. See dplyr::select and tidyselect::select_helpers for details.

See also

Examples

# For demonstration only, since ICA won't converge
# Suppressing an important warning:
suppressWarnings(data_faces_10_trials %>%
  eeg_ica(-EOGH, -EOGV, -M1, -M2, .method = fast_ICA, .config = list(maxit = 10))) %>%
  eeg_ica_keep(-ICA1)
#> # ICA is being done using fast_ICA...
#> # ICA finished in 9 iterations. With a tolerance of 0.863658. (Maximum expected tolerance: 1e-06.)
#> # ICA took 1.95 secs
#> # EEG data:
#> 
#> # Signal table:
#>        .id .sample       Fp1       Fpz        Fp2        F7         F3
#>     1:   1   15000 -41.55177 -12.34044 -11.479909 -1.154271 -18.663160
#>     2:   1   15001 -38.34803 -19.89935 -10.474332  2.409413 -16.290863
#>     3:   1   15002 -34.17843 -23.23791  -6.899619  6.683236 -11.469943
#>     4:   1   15003 -32.02068 -22.76313  -3.797027 11.102965 -10.254589
#>     5:   1   15004 -32.84556 -20.85845  -1.679441 12.555004  -9.986915
#>    ---                                                                
#> 43325:   1   58324 -20.18384 -14.40326 -10.232825 -1.186729  -8.802548
#> 43326:   1   58325 -20.00285 -13.52485  -9.594756 -1.445169  -7.937759
#> 43327:   1   58326 -18.93813 -13.88643 -10.136325 -1.569261  -7.947565
#> 43328:   1   58327 -18.38915 -15.04560 -11.661365 -3.310127  -7.974796
#> 43329:   1   58328 -17.64114 -15.61484 -12.404674 -4.292785  -7.845066
#>                Fz         F4        F8        FC5          FC1         FC2
#>     1:  -6.694997 -5.3587816  9.232065  1.2476285 -17.95894243  -6.5439535
#>     2:  -9.136874 -7.2320180 10.226788  2.2819662 -15.36088962  -8.9088355
#>     3: -10.228896 -7.6666159  9.815068 -2.7103966 -10.46040733 -10.6660473
#>     4: -10.674997 -7.9005497  9.464712 -6.0187260  -9.89913731 -12.3407636
#>     5: -10.145313 -7.0603373 10.381191 -2.7763079 -10.97653896 -12.2829799
#>    ---                                                                    
#> 43325:  -7.925756 -4.5663952 -9.721773 -7.9841711  -1.00803039  -2.5760820
#> 43326:  -6.954232 -3.0589571 -9.193802 -9.5184556  -0.16419547  -1.3277711
#> 43327:  -6.434799 -2.0471016 -8.147936 -8.0343799   0.09546457  -0.8875517
#> 43328:  -6.181899 -1.1624159 -7.271946 -4.0661379   0.30312677  -0.4309207
#> 43329:  -5.884319 -0.5664237 -6.373051  0.9913301  -0.20902399  -0.1892064
#>              FC6         M1          T7         C3           Cz         C4
#>     1: 11.282339 -1.8668146 -16.4058047 -17.916217 -21.34119198  1.9330840
#>     2: -2.215138 -3.2960243  -8.6422289 -14.236119 -18.43185343 -0.3233588
#>     3: -9.682181 -3.7450671  -4.1074325  -5.655800 -14.88131499 -1.4903017
#>     4: -9.735252 -3.7946916  -8.1341500  -1.665289 -16.32433259 -2.1053437
#>     5: -5.273119 -3.2719884 -13.4946805  -2.800548 -19.02947404 -2.0786687
#>    ---                                                                    
#> 43325: -8.658844  0.8050708   0.4788326   1.039157  -3.31018337 -7.4993707
#> 43326: -6.910211  0.9832335   0.7573455   2.278589  -1.79990800 -5.8042494
#> 43327: -5.212926  1.2051768   1.0644705   2.421607  -0.73580724 -4.6594645
#> 43328: -3.281097  1.4103985   1.5013841   1.423897  -0.04167516 -3.8103833
#> 43329: -2.393035  1.1627738   1.5901038  -1.069263  -0.53245821 -3.3186731
#>               T8         M2        CP5        CP1        CP2        CP6
#>     1:  3.715371  1.8668146 -6.3991310 -28.056393 -10.270281  -3.883333
#>     2: 14.897235  3.2960243 -5.2086968 -23.220972  -8.626375   4.003721
#>     3: 24.215521  3.7450671 -0.5714621 -16.513941  -6.077231  11.286855
#>     4: 26.474599  3.7946916  3.9655664 -16.097134  -6.722799  12.483585
#>     5: 22.873994  3.2719884  6.0243852 -18.591615  -8.864493   8.763486
#>    ---                                                                 
#> 43325: -7.930700 -0.8050708  4.3575710   2.373151  -7.259431 -12.513446
#> 43326: -7.113160 -0.9832335  4.2947479   3.309556  -5.728292 -10.109401
#> 43327: -2.964021 -1.2051768  3.7357260   3.643422  -4.671715  -7.665981
#> 43328: -1.006426 -1.4103985  3.2903825   4.292440  -4.006717  -4.766592
#> 43329: -0.762234 -1.1627738  2.3798967   3.601423  -4.414848  -3.104586
#>                P7           P3          Pz         P4         P8         POz
#>     1: -2.4870646 -12.45340977 -19.2175174 -11.077772 -18.474206 -20.9080713
#>     2: -0.9615085 -10.83608589 -16.0721969  -6.962400 -10.098915 -16.5090411
#>     3:  1.8156595  -6.98722845 -12.2251191  -2.855115  -2.631139 -12.1269383
#>     4:  3.5707762  -5.85883147 -13.7455506  -3.548441  -2.606201 -13.2080003
#>     5:  3.9046053  -6.84357497 -16.9369898  -6.878236  -7.188727 -16.8309759
#>    ---                                                                      
#> 43325:  8.1097311   2.23357868  -2.7781400  -9.022425 -10.214366  -3.5511564
#> 43326:  7.1451154   2.08786411  -1.8606857  -7.540730  -8.708988  -2.5901059
#> 43327:  5.0165121   1.33923441  -1.0987279  -5.724186  -5.832742  -1.4209683
#> 43328:  3.6754622   1.11618570   0.1487296  -3.483917  -2.646955   0.2989573
#> 43329:  2.8776821  -0.07198172  -0.1559751  -2.666364  -1.295383   0.4067311
#>                 O1           Oz          O2       EOGV       EOGH
#>     1: -19.5704747 -22.55997785 -21.4112541 -205.85693 164.320831
#>     2: -14.3624877 -16.27612534 -14.7499274 -214.15259 158.389755
#>     3:  -9.1174749 -10.69768423  -9.0707173 -220.57965 157.559784
#>     4:  -8.2341462 -11.21936779  -9.8982453 -218.24934 162.165161
#>     5: -11.3239908 -16.56303859 -15.3936998 -217.16670 168.559723
#>    ---                                                           
#> 43325:   2.3422476  -4.26503338  -7.3075605   18.41424   6.434383
#> 43326:   2.0570941  -3.26667819  -5.7686898   19.22052   7.837689
#> 43327:   1.0856791  -1.80492566  -2.9308017   19.74028   8.659687
#> 43328:   0.8561164   0.07461558   0.2777538   19.20309   8.609533
#> 43329:   0.7975315   0.70924032   1.2777655   18.72752   8.829977
#> 
#> ## Events table:
#>      .id        .type .description .initial .final .channel
#>   1:   1 Bad Interval  Bad Min-Max    15000  15128      FC6
#>   2:   1 Bad Interval  Bad Min-Max    15000  15146       F8
#>   3:   1 Bad Interval  Bad Min-Max    15000  15142      Fp1
#>   4:   1 Bad Interval  Bad Min-Max    15000  15132      Fp2
#>   5:   1 Bad Interval  Bad Min-Max    15000  15140      Fpz
#>  ---                                                       
#> 540:   1 Bad Interval  Bad Min-Max    55058  55400      Fpz
#> 541:   1 Bad Interval  Bad Min-Max    55077  55314       F8
#> 542:   1     Stimulus         s130    56380  56380     <NA>
#> 543:   1     Stimulus          s70    56750  56750     <NA>
#> 544:   1     Stimulus         s130    58328  58328     <NA>
#> 
#> # ICA:
#> # Component_names:ICA2, ICA3, ICA4, ICA5, ICA6, ICA7, ICA8, ICA9, ICA10, ICA11, ICA12, ICA13, ICA14, ICA15, ICA16, ICA17, ICA18, ICA19, ICA20, ICA21, ICA22, ICA23, ICA24, ICA25, ICA26, ICA27, ICA28, ICA29, ICA30
#> # Channels_used: Fp1, Fpz, Fp2, F7, F3, Fz, F4, F8, FC5, FC1, FC2, FC6, T7, C3, Cz, C4, T8, CP5, CP1, CP2, CP6, P7, P3, Pz, P4, P8, POz, O1, Oz, O2
#> 
#> # Segments table:
#>    .id .recording segment
#> 1:   1 faces.vhdr       1