Select independent components (or sources) to keep.

This function will transform the channels according to the independent components that are kept or removed.

eeg_ica_keep(.data, ...)

Arguments

.data

An eeg_ica_lst object

...

Components to keep from the mixing matrix of the ICA transformation. See dplyr::select and tidyselect::select_helpers for details.

See also

Examples

# For demonstration only, since ICA won't converge
# Suppressing an important warning:
suppressWarnings(data_faces_10_trials %>%
  eeg_ica(-EOGH, -EOGV, -M1, -M2, .method = fast_ICA, .config = list(maxit = 10))) %>%
  eeg_ica_keep(-ICA1)
#> # ICA is being done using fast_ICA...
#> # ICA finished in 9 iterations. With a tolerance of 0.965564. (Maximum expected tolerance: 1e-06.)
#> # ICA took 1.29 secs
#> # EEG data:
#> 
#> # Signal table:
#>        .id .sample       Fp1        Fpz         Fp2         F7         F3
#>     1:   1   15000 -24.32677  -3.725814  -6.0387069  1.3833058  -6.328811
#>     2:   1   15001 -25.82722 -11.572234  -5.4347694  4.5961826  -8.391252
#>     3:   1   15002 -25.99895 -16.899279  -3.1396034  8.4068084  -6.936190
#>     4:   1   15003 -26.72896 -17.710712  -1.0670810 10.1926887  -5.910830
#>     5:   1   15004 -25.73971 -16.322936   0.8865939 11.8032592  -3.699486
#>    ---                                                                   
#> 43325:   1   58324 -20.08211 -12.906906  -9.3484012 -0.2132441 -10.048773
#> 43326:   1   58325 -19.62417 -12.680408  -9.0432757 -0.5008061  -8.661892
#> 43327:   1   58326 -20.13531 -13.217725  -9.8563836 -1.7907985  -9.225930
#> 43328:   1   58327 -19.62480 -14.028919 -11.1207812 -2.5359524 -10.311108
#> 43329:   1   58328 -18.84423 -14.429777 -11.7115464 -2.7953581 -10.841019
#>               Fz        F4         F8         FC5       FC1        FC2
#>     1:  3.537160  3.694760  0.1952318  8.64387893 -2.037024  0.6090740
#>     2: -1.549230 -0.606540  4.3525601  7.94968562 -5.068172 -3.3079627
#>     3: -5.272717 -3.399177  6.3648718  1.18100472 -4.488230 -6.9348802
#>     4: -6.260486 -3.807135  6.1138110 -3.90108829 -3.751873 -8.2696066
#>     5: -4.829847 -2.135658  5.6770417 -0.08274567 -2.448460 -7.8208474
#>    ---                                                                
#> 43325: -8.150844 -4.920324 -8.7817936 -7.62046839 -2.780320 -2.7249905
#> 43326: -7.096173 -3.312169 -8.6327791 -9.10447263 -1.266090 -1.5621317
#> 43327: -6.926100 -2.491448 -7.2446462 -8.45016071 -1.466862 -0.9646431
#> 43328: -7.149723 -2.177851 -5.6248367 -4.20050511 -2.749765 -1.0912058
#> 43329: -7.179500 -1.978617 -4.2487606  1.05576344 -4.235060 -1.2769288
#>               FC6         M1           T7          C3         Cz        C4
#>     1: 13.9655683 -1.8668146 -12.56556892 -3.86106384 -10.197713  8.765819
#>     2:  0.2380589 -3.2960243  -6.27150631 -4.73100431 -10.708242  4.964729
#>     3: -7.9643278 -3.7450671  -2.75874688  0.07051037 -10.135827  2.000182
#>     4: -7.1326440 -3.7946916  -7.05097805  4.25452325 -11.258079  1.927582
#>     5: -2.8136682 -3.2719884 -11.73780652  4.92125990 -12.664109  2.351719
#>    ---                                                                    
#> 43325: -8.6617109  0.8050708   0.06742909 -0.26479966  -4.258918 -7.767266
#> 43326: -7.1115054  0.9832335   0.56156453  1.41340462  -2.475591 -6.129220
#> 43327: -4.9207991  1.2051768   0.58362688  1.28769803  -1.501806 -4.744520
#> 43328: -3.4655754  1.4103985   0.77017644 -0.95560101  -1.820512 -4.548098
#> 43329: -2.9456073  1.1627738   0.70749872 -4.28450996  -3.002485 -4.534596
#>                 T8         M2         CP5          CP1       CP2        CP6
#>     1: -24.7084480  1.8668146  3.26874328 -13.38828114 -7.790748   5.501269
#>     2:  -4.2922590  3.2960243  1.19830927 -13.15862479 -6.980020  10.768211
#>     3:  12.7539681  3.7450671  3.20116898 -10.42695224 -5.183345  15.436515
#>     4:  13.2846257  3.7946916  8.02988052  -9.12785925 -4.859599  18.329377
#>     5:   6.0211685  3.2719884 11.41454686  -9.87694259 -6.695086  15.469475
#>    ---                                                                     
#> 43325:  -4.9535263 -0.8050708  3.32211996   0.90965389 -7.740837 -13.507650
#> 43326:  -4.9624549 -0.9832335  3.59883378   2.22780917 -6.173099 -11.044989
#> 43327:  -0.8176998 -1.2051768  2.89515994   2.63786427 -4.758773  -7.976540
#> 43328:   4.2342785 -1.4103985  1.50022442   1.73925559 -4.724173  -6.500510
#> 43329:   6.5592586 -1.1627738 -0.04183755  -0.01482846 -5.570965  -5.838416
#>              P7         P3         Pz          P4           P8         POz
#>     1: 2.450489 -3.2100794 -10.190881 -6.92676207 -13.28583511 -14.4849320
#>     2: 1.974979 -4.3752894  -9.867948 -3.55585181  -6.09205545 -11.6664392
#>     3: 3.329593 -3.0727288  -8.531987 -0.61799805  -0.01407185  -9.0541539
#>     4: 5.765601 -0.8359545  -8.722765  0.05726699   0.76972033  -8.8804746
#>     5: 6.995208 -0.8075495 -10.862687 -3.19379381  -3.51906077 -12.0553228
#>    ---                                                                    
#> 43325: 7.187073  1.2437784  -3.881820 -9.33562592 -10.55241876  -4.1330352
#> 43326: 6.496122  1.2607618  -2.775755 -8.01524610  -9.09291734  -3.2116317
#> 43327: 4.445452  0.8356268  -1.617542 -5.53486561  -5.86854103  -1.5043411
#> 43328: 2.326376 -0.5761012  -1.659320 -4.09989318  -3.32299518  -0.7391031
#> 43329: 1.016431 -2.5874749  -2.855063 -3.87133704  -2.44000515  -1.3142729
#>                  O1          Oz           O2       EOGV       EOGH
#>     1: -11.74594527 -15.3151185 -17.25616422 -205.85693 164.320831
#>     2:  -8.44785182 -10.8148253 -11.02186761 -214.15259 158.389755
#>     3:  -5.27576957  -7.1554994  -6.41386538 -220.57965 157.559784
#>     4:  -3.87720676  -7.1086456  -6.46377757 -218.24934 162.165161
#>     5:  -6.39701052 -11.9177998 -12.06971559 -217.16670 168.559723
#>    ---                                                            
#> 43325:   1.91137649  -4.7040772  -7.26504815   18.41424   6.434383
#> 43326:   1.62090445  -3.7043351  -5.96745335   19.22052   7.837689
#> 43327:   0.85822457  -2.0124748  -2.63931553   19.74028   8.659687
#> 43328:  -0.09461483  -0.8424958   0.09981872   19.20309   8.609533
#> 43329:  -0.67117425  -0.7175677   0.73134558   18.72752   8.829977
#> 
#> ## Events table:
#>      .id        .type .description .initial .final .channel
#>   1:   1 Bad Interval  Bad Min-Max    15000  15128      FC6
#>   2:   1 Bad Interval  Bad Min-Max    15000  15146       F8
#>   3:   1 Bad Interval  Bad Min-Max    15000  15142      Fp1
#>   4:   1 Bad Interval  Bad Min-Max    15000  15132      Fp2
#>   5:   1 Bad Interval  Bad Min-Max    15000  15140      Fpz
#>  ---                                                       
#> 540:   1 Bad Interval  Bad Min-Max    55058  55400      Fpz
#> 541:   1 Bad Interval  Bad Min-Max    55077  55314       F8
#> 542:   1     Stimulus         s130    56380  56380     <NA>
#> 543:   1     Stimulus          s70    56750  56750     <NA>
#> 544:   1     Stimulus         s130    58328  58328     <NA>
#> 
#> # ICA:
#> # Component_names:ICA2, ICA3, ICA4, ICA5, ICA6, ICA7, ICA8, ICA9, ICA10, ICA11, ICA12, ICA13, ICA14, ICA15, ICA16, ICA17, ICA18, ICA19, ICA20, ICA21, ICA22, ICA23, ICA24, ICA25, ICA26, ICA27, ICA28, ICA29, ICA30
#> # Channels_used: Fp1, Fpz, Fp2, F7, F3, Fz, F4, F8, FC5, FC1, FC2, FC6, T7, C3, Cz, C4, T8, CP5, CP1, CP2, CP6, P7, P3, Pz, P4, P8, POz, O1, Oz, O2
#> 
#> # Segments table:
#>    .id .recording segment
#> 1:   1 faces.vhdr       1