Segment an eeg_lst.

Subdivide of the EEG into different segments or epochs. When there is no segmentation, the eeg_lst contain one segment. (Fieldtrip calls the segment "trials".) The limits of segment are inclusive: If, for example, lim = c(0,0), the segment would contain only sample 1.

eeg_segment(x, ...)

# S3 method for eeg_lst
eeg_segment(x, ..., lim = c(-0.5, 0.5), end,
  unit = "seconds", recording_col = "recording")

Arguments

x
An eeg_lst object.

...
Description of the event.

lim
Vector indicating the time before and after the event. Or dataframe with two columns, with nrow=total number of segments

end
Description of the event that indicates the end of the segment, if this is used, lim is ignored.

unit
Unit

recording_col
Column in the segments table indicating to which recording or file each segment belongs.

Value

An eeg_lst.